latestNewsCard removal
