+++ src/features/blog/components/newsCard/_spec.js
@@ -5,13 +5,17 @@ import NewsCard from '.'
 describe(NewsCard.name, () => {
   it('renders with default values', () => {
     const title = 'Here is some news!'
-    const filterType = {
+    const category = {
       hexColour: '#304242',
       title: 'Amazing News!!',
     }
-    const date = '2019-03-28T20:30+00:00'
+    const datePublished = '2019-03-28T20:30+00:00'
     const wrapper = shallow(
-      <NewsCard filterType={filterType} title={title} date={date} />
+      <NewsCard
+        category={category}
+        title={title}
+        datePublished={datePublished}
+      />
     )
     expect(wrapper).toMatchSnapshot()
   })
